+++ CHANGELOG.md
@@ -35,6 +35,7 @@ We also added 3 precompiled packages for browsers and Node:
 - **[BREAKING]** renamed `makeUnclosable` into `asUnclosable`.
 - **[BREAKING]** renamed the old `delay` into `wait`.
 - **[BREAKING]** renamed `unwrap` into `unwrapPromise`.
+- **[BREAKING]** renamed `pullItems` and `pushItems` into `pullArray` and `pushArray`.
 
 ### Removed